deleted extra unused files from repository
this project will be documented using NaturalDocs
i haved decided to only document files in src or
include files.
the project directory was put outside as well as the docs
directory generated by naturalDocs.

adding bash file for the naturalDocs to codeBlocks
buildProccess. the name of the file is Update_Docs.

 Changes to be committed:
	deleted:    include/boundingBox.hpp
	modified:   include/bullet.h
	deleted:    include/textureId.h
	deleted:    include/textureManger.h
	deleted:    include/visGameObj.h
	modified:   shooterGame.cbp
	deleted:    src/boundingBox.cpp
	deleted:    src/visGameObj.cpp